requirements.txt
Flask==3.0.3
Pillow==10.3.0
flask-limiter==3.5.0
gunicorn==22.0.0
requests==2.32.3
psutil>=5.9.0
numpy>=2.2
Werkzeug>=3.0.0